fix: fix and move event helpers
